Add andR, orR, and xorR methods
These are convenience methods for m.reduce(<op>, <val>).  It's a common
operation and it's significantly less verbose to just call one of these
methods rather than having to do the full m.reduce and operator.<op>
style

To avoid a circular dependency between reduce.py and bits.py, I moved
the reduce logic into bits.  This seems reasonable since the other bits
related primitives are already housed there.
