+++ magma/__init__.py
@@ -100,6 +100,6 @@ from .syntax.inline_combinational import inline_combinational
 from .syntax.coroutine import coroutine
 
 from magma.primitives import (LUT, Mux, mux, Register, get_slice, set_slice,
-                              slice, reduce, Memory)
+                              slice, Memory)
 
 from magma.types import BitPattern
