+++ magma/primitives/__init__.py
@@ -2,5 +2,4 @@ from magma.primitives.lut import LUT
 from magma.primitives.memory import Memory
 from magma.primitives.mux import Mux, mux
 from magma.primitives.register import Register
-from magma.primitives.reduce import reduce
 from magma.primitives.slice import get_slice, set_slice, slice
