+++ tests/test_primitives/gold/test_reduce_and_.v
@@ -1,7 +1,9 @@
 module test_reduce_and_ (
     input [4:0] I,
-    output O
+    output O0,
+    output O1
 );
-assign O = & I;
+assign O0 = & I;
+assign O1 = & I;
 endmodule
